Refactor currency conversion logic and support CSSBuy frontend
The code changes refactor the currency conversion logic in the `content.js` file. The previous implementation only supported the Superbuy frontend, but now it also supports the CSSBuy frontend. The changes include modifying the selector for the goods text element and updating the currency conversion algorithm. This allows the price to be converted correctly when flicking through a product's variations. The code also adds support for detecting and converting prices displayed in a `<p>` tag with a nested `<span>` element. Overall, these changes improve the functionality and flexibility of the currency conversion feature. Currently, switching between the variations of products only works on superbuy and not cssbuy due to the way cssbuy works. this will likely not be fixed anytime soon.